Revert changes to HD as the HD team is going to take care of the update.

+++ com.unity.render-pipelines.high-definition/Editor/AssetProcessors/FBXArnoldSurfaceMaterialDescriptionPreprocessor.cs
@@ -102,8 +102,8 @@ namespace UnityEditor.Rendering.HighDefinition
                     material.SetFloat("_OPACITY", opacity);
                 }
                 
-                material.SetFloat("_SrcBlend", 1.0f);
-                material.SetFloat("_DstBlend", 10.0f);
+                material.SetInt("_SrcBlend", 1);
+                material.SetInt("_DstBlend", 10);
                 material.SetFloat("_BlendMode", (float)BlendMode.Alpha);
                 material.EnableKeyword("_ALPHAPREMULTIPLY_ON");
                 material.EnableKeyword("_SURFACE_TYPE_TRANSPARENT");
@@ -114,8 +114,8 @@ namespace UnityEditor.Rendering.HighDefinition
             else
             {
                 material.EnableKeyword("_DOUBLESIDED_ON");
-                material.SetFloat("_CullMode", 0.0f);
-                material.SetFloat("_CullModeForward", 0.0f);
+                material.SetInt("_CullMode", 0);
+                material.SetInt("_CullModeForward", 0);
                 material.doubleSidedGI = true;
             }
 
@@ -207,8 +207,8 @@ namespace UnityEditor.Rendering.HighDefinition
                     material.SetFloat("_OPACITY", opacity);
                 }
 
-                material.SetFloat("_SrcBlend", 1.0f);
-                material.SetFloat("_DstBlend", 10.0f);
+                material.SetInt("_SrcBlend", 1);
+                material.SetInt("_DstBlend", 10);
                 material.SetFloat("_BlendMode", (float)BlendMode.Alpha);
                 material.EnableKeyword("_ALPHAPREMULTIPLY_ON");
                 material.EnableKeyword("_SURFACE_TYPE_TRANSPARENT");
@@ -219,8 +219,8 @@ namespace UnityEditor.Rendering.HighDefinition
             else
             {
                 material.EnableKeyword("_DOUBLESIDED_ON");
-                material.SetFloat("_CullMode", 0.0f);
-                material.SetFloat("_CullModeForward", 0.0f);
+                material.SetInt("_CullMode", 0);
+                material.SetInt("_CullModeForward", 0);
                 material.doubleSidedGI = true;
             }
