+++ com.unity.render-pipelines.high-definition/Editor/Material/Lit/LitGUI.cs
@@ -84,8 +84,8 @@ namespace UnityEditor.Rendering.HighDefinition
         {
             BaseLitGUI.SetupBaseLitKeywords(material);
             BaseLitGUI.SetupBaseLitMaterialPass(material);
-            bool receiveSSR = material.GetSurfaceType() == SurfaceType.Opaque ? (material.HasProperty(kReceivesSSR) ? (int)material.GetFloat(kReceivesSSR) != 0 : false)
-                                                : (material.HasProperty(kReceivesSSRTransparent) ? (int)material.GetFloat(kReceivesSSRTransparent) != 0 : false);
+            bool receiveSSR = material.GetSurfaceType() == SurfaceType.Opaque ? (material.HasProperty(kReceivesSSR) ? material.GetInt(kReceivesSSR) != 0 : false)
+                                                : (material.HasProperty(kReceivesSSRTransparent) ? material.GetInt(kReceivesSSRTransparent) != 0 : false);
             BaseLitGUI.SetupStencil(material, receiveSSR, material.GetMaterialId() == MaterialId.LitSSS);
 
             if (material.HasProperty(kNormalMapSpace))
@@ -131,7 +131,7 @@ namespace UnityEditor.Rendering.HighDefinition
                 // _ENABLESPECULAROCCLUSION is deprecated
                 CoreUtils.SetKeyword(material, "_ENABLESPECULAROCCLUSION", false);
 
-                int specOcclusionMode = (int)material.GetFloat(kSpecularOcclusionMode);
+                int specOcclusionMode = material.GetInt(kSpecularOcclusionMode);
                 CoreUtils.SetKeyword(material, "_SPECULAR_OCCLUSION_NONE", specOcclusionMode == 0);
                 CoreUtils.SetKeyword(material, "_SPECULAR_OCCLUSION_FROM_BENT_NORMAL_MAP", specOcclusionMode == 2);
             }
@@ -198,7 +198,7 @@ namespace UnityEditor.Rendering.HighDefinition
 
             if (material.HasProperty(kAddPrecomputedVelocity))
             {
-                CoreUtils.SetKeyword(material, "_ADD_PRECOMPUTED_VELOCITY", (int)material.GetFloat(kAddPrecomputedVelocity) != 0);
+                CoreUtils.SetKeyword(material, "_ADD_PRECOMPUTED_VELOCITY", material.GetInt(kAddPrecomputedVelocity) != 0);
             }
         }
     }
