+++ com.unity.render-pipelines.high-definition/Runtime/Material/Decal/DecalSystem.cs
@@ -862,7 +862,7 @@ namespace UnityEngine.Rendering.HighDefinition
                 {
                     if (m_IsHDRenderPipelineDecal)
                     {
-                        return (int)this.m_Material.GetFloat("_DrawOrder");
+                        return this.m_Material.GetInt("_DrawOrder");
                     }
                     else
                     {
